Fix components

+++ frontend/app/src/main/java/com/example/smartstorageorganizer/adapters/CategoryAdapter.java
@@ -22,6 +22,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.example.smartstorageorganizer.HomeActivity;
+import com.example.smartstorageorganizer.MyAmplifyApp;
 import com.example.smartstorageorganizer.R;
 import com.example.smartstorageorganizer.UncategorizedItemsActivity;
 import com.example.smartstorageorganizer.ViewItemActivity;
@@ -30,9 +31,14 @@ import com.example.smartstorageorganizer.ui.home.HomeFragment;
 import com.example.smartstorageorganizer.utils.OperationCallback;
 import com.example.smartstorageorganizer.utils.Utils;
 import com.google.android.material.bottomsheet.BottomSheetDialog;
+import com.google.firebase.firestore.FieldValue;
+import com.google.firebase.firestore.FirebaseFirestore;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
+import java.util.concurrent.CompletableFuture;
 
 public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.ViewHolder> {
 
@@ -41,10 +47,13 @@ public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.ViewHo
     private String organizationID;
     private HomeFragment homeFragment;
 
+    private MyAmplifyApp app;
+
     public CategoryAdapter(Context context, List<CategoryModel> categoryModelList, HomeFragment homeFragment) {
         this.context = context;
         this.categoryModelList = categoryModelList;
         this.homeFragment = homeFragment;
+        app = (MyAmplifyApp) context.getApplicationContext();
     }
 
     @NonNull
@@ -158,8 +167,12 @@ public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.ViewHo
             updateButton.setOnClickListener(v -> {
                 String newName = editCategoryName.getText().toString().trim();
                 if (!newName.isEmpty() && !newName.equals(category.getCategoryName())) {
-                    // Update the category name (you need to implement the updateCategoryName method)
-                    updateCategoryName(Integer.parseInt(category.getCategoryID()), newName);
+                    if(Objects.equals(app.getUserRole(), "Manager") || Objects.equals(app.getUserRole(), "Admin")){
+                        updateCategoryName(Integer.parseInt(category.getCategoryID()), newName);
+                    }
+                    else if(Objects.equals(app.getUserRole(), "normalUser")){
+                        sendRequestToModifyCategory(category.getCategoryID(), category.getCategoryName(), newName);
+                    }
                     editDialog.dismiss();
                 }
             });
@@ -193,7 +206,12 @@ public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.ViewHo
                 .setMessage("Do you want to delete the " + category.getCategoryName() + " category?")
                 .setPositiveButton("Yes", (dialog, which) -> {
                     Log.i("Adapter", "Yes clicked.");
-                    deleteCategory(Integer.parseInt(category.getCategoryID()));
+                    if(Objects.equals(app.getUserRole(), "Manager") || Objects.equals(app.getUserRole(), "Admin")){
+                        deleteCategory(Integer.parseInt(category.getCategoryID()));
+                    }
+                    else if(Objects.equals(app.getUserRole(), "normalUser")){
+                        sendRequestToDeleteCategory(category.getCategoryID(), category.getCategoryName());
+                    }
                 })
                 .setNegativeButton("No", (dialog, which) -> {
                     // Dismiss the dialog
@@ -203,12 +221,171 @@ public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.ViewHo
                 .show();
     }
 
+    private void navigateToHome() {
+        Intent intent = new Intent(context, ViewItemActivity.class);
+        context.startActivity(intent);
+    }
+
+    public void setOrganizationId(String organizationId) {
+        this.organizationID = organizationId;
+    }
+
+    public void showRequestDialog() {
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+        LayoutInflater inflater = LayoutInflater.from(context);
+        View dialogView = inflater.inflate(R.layout.send_request_popup, null);
+
+        builder.setView(dialogView);
+        AlertDialog alertDialog = builder.create();
+        Button closeButton = dialogView.findViewById(R.id.finishButton);
+
+        closeButton.setOnClickListener(v -> {
+            alertDialog.dismiss();
+
+//            Intent intent = new Intent(context, HomeActivity.class);
+//            context.startActivity(intent);
+//
+//            if (context instanceof Activity) {
+//                ((Activity) context).finish();
+//            }
+        });
+
+        alertDialog.setCanceledOnTouchOutside(false);
+        alertDialog.show();
+    }
+
+
+    public CompletableFuture<Boolean> sendRequestToDeleteCategory(String id, String categoryName) {
+        ProgressDialog progressDialog = new ProgressDialog(context);
+        progressDialog.setMessage("Sending Request To Delete a Category...");
+        progressDialog.setCancelable(false);
+        progressDialog.show();
+        CompletableFuture<Boolean> future = new CompletableFuture<>();
+
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+
+        // Create a map for the unit request
+        Map<String, Object> unitRequest = new HashMap<>();
+        unitRequest.put("id", id);
+        unitRequest.put("categoryName", categoryName);
+        unitRequest.put("userEmail", app.getEmail());
+        unitRequest.put("requestType", "Delete Category");
+        unitRequest.put("organizationId", app.getOrganizationID());
+        unitRequest.put("status", "pending");  // Initially set to pending
+        unitRequest.put("requestDate", FieldValue.serverTimestamp()); // Store request date and time
+
+        // Store the request in Firestore
+        db.collection("category_requests")
+                .add(unitRequest)
+                .addOnSuccessListener(documentReference -> {
+                    // Get the unique document ID
+                    String documentId = documentReference.getId();
+
+                    // Update the document to include the document ID or use it as a unique ID
+                    db.collection("category_requests").document(documentId)
+                            .update("documentId", documentId) // Store documentId within the document itself
+                            .addOnSuccessListener(aVoid -> {
+                                progressDialog.dismiss();
+                                showRequestDialog();
+                                Log.i("Firestore", "Request stored successfully with documentId: " + documentId);
+                                future.complete(true);
+                            })
+                            .addOnFailureListener(e -> {
+                                progressDialog.dismiss();
+                                Log.e("Firestore", "Error updating documentId", e);
+                                future.complete(false);
+                            });
+                })
+                .addOnFailureListener(e -> {
+                    progressDialog.dismiss();
+                    Log.e("Firestore", "Error storing request", e);
+                    future.complete(false);
+                });
+
+        return future;
+    }
+
+    public void sendRequestToModifyCategory(String id, String currentCategoryName, String newCategoryName) {
+        ProgressDialog progressDialog = new ProgressDialog(context);
+        progressDialog.setMessage("Sending Request to Update Category Name...");
+        progressDialog.setCancelable(false);
+        progressDialog.show();
+
+        CompletableFuture<Boolean> future = new CompletableFuture<>();
+
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+
+        // Create a map for the unit request
+        Map<String, Object> unitRequest = new HashMap<>();
+        unitRequest.put("id", id);
+        unitRequest.put("currentCategoryName", currentCategoryName);
+        unitRequest.put("newCategoryName", newCategoryName);
+        unitRequest.put("userEmail", app.getEmail());
+        unitRequest.put("requestType", "Modify Category");
+        unitRequest.put("organizationId", app.getOrganizationID());
+        unitRequest.put("status", "pending");  // Initially set to pending
+        unitRequest.put("requestDate", FieldValue.serverTimestamp()); // Store request date and time
+
+        // Store the request in Firestore
+        db.collection("category_requests")
+                .add(unitRequest)
+                .addOnSuccessListener(documentReference -> {
+                    // Get the unique document ID
+                    String documentId = documentReference.getId();
+
+                    // Update the document to include the document ID or use it as a unique ID
+                    db.collection("category_requests").document(documentId)
+                            .update("documentId", documentId) // Store documentId within the document itself
+                            .addOnSuccessListener(aVoid -> {
+                                progressDialog.dismiss();
+                                showRequestDialog();
+                                Log.i("Firestore", "Request stored successfully with documentId: " + documentId);
+                                future.complete(true);
+                            })
+                            .addOnFailureListener(e -> {
+                                progressDialog.dismiss();
+                                Log.e("Firestore", "Error updating documentId", e);
+                                future.complete(false);
+                            });
+                })
+                .addOnFailureListener(e -> {
+                    progressDialog.dismiss();
+                    Log.e("Firestore", "Error storing request", e);
+                    future.complete(false);
+                });
+
+    }
+
+    private void updateCategoryName(int categoryId, String newName) {
+        ProgressDialog progressDialog = new ProgressDialog(context);
+        progressDialog.setMessage("Updating category name...");
+        progressDialog.setCancelable(false);
+        progressDialog.show();
+        Utils.modifyCategoryName(categoryId, newName, app.getEmail(), app.getOrganizationID(), (Activity) context, new OperationCallback<Boolean>(){
+            @Override
+            public void onSuccess(Boolean result) {
+                progressDialog.dismiss();
+                if (Boolean.TRUE.equals(result)) {
+                    Toast.makeText(context, "Category Name Changed Successfully.", Toast.LENGTH_SHORT).show();
+                    Intent intent = new Intent(context, HomeActivity.class);
+                    context.startActivity(intent);
+                    ((Activity) context).finish();
+                }
+            }
+            @Override
+            public void onFailure(String error) {
+                progressDialog.dismiss();
+                Toast.makeText(context, "Failed to Modify Category Name.", Toast.LENGTH_SHORT).show();
+            }
+        });
+    }
+
     private void deleteCategory(int id) {
         ProgressDialog progressDialog = new ProgressDialog(context);
         progressDialog.setMessage("Deleting category...");
         progressDialog.setCancelable(false);
         progressDialog.show();
-        Utils.deleteCategory(id, "NULL", (Activity) context, new OperationCallback<Boolean>() {
+        Utils.deleteCategory(id, "NULL", app.getEmail(), (Activity) context, new OperationCallback<Boolean>() {
             @Override
             public void onSuccess(Boolean result) {
                 if (Boolean.TRUE.equals(result)) {
@@ -247,38 +424,5 @@ public class CategoryAdapter extends RecyclerView.Adapter<CategoryAdapter.ViewHo
                 Toast.makeText(context, "Failed to Delete Category.", Toast.LENGTH_SHORT).show();
             }
         });
-    }
-
-    private void updateCategoryName(int categoryId, String newName) {
-        ProgressDialog progressDialog = new ProgressDialog(context);
-        progressDialog.setMessage("Updating category name...");
-        progressDialog.setCancelable(false);
-        progressDialog.show();
-        Utils.modifyCategoryName(categoryId, newName, (Activity) context, new OperationCallback<Boolean>(){
-            @Override
-            public void onSuccess(Boolean result) {
-                progressDialog.dismiss();
-                if (Boolean.TRUE.equals(result)) {
-                    Toast.makeText(context, "Category Name Changed Successfully.", Toast.LENGTH_SHORT).show();
-                    Intent intent = new Intent(context, HomeActivity.class);
-                    context.startActivity(intent);
-                    ((Activity) context).finish();
-                }
-            }
-            @Override
-            public void onFailure(String error) {
-                progressDialog.dismiss();
-                Toast.makeText(context, "Failed to Modify Category Name.", Toast.LENGTH_SHORT).show();
-            }
-        });
-    }
-
-    private void navigateToHome() {
-        Intent intent = new Intent(context, ViewItemActivity.class);
-        context.startActivity(intent);
-    }
-
-    public void setOrganizationId(String organizationId) {
-        this.organizationID = organizationId;
     }
 }
